requirements.txt
RPi.GPIO
spidev
requests
PyYAML